procrustes and pca warnings removed

+++ ImageInLib/ImageInLib/src/procrustrates_dist.h
@@ -15,7 +15,7 @@ typedef struct {
 // Generalized Procrustes Analysis - Returns the Mean Shape
 void genProcMeanShape(dataType ** dtaMnShp, Shapes *shapes, size_t height, size_t length, size_t width, size_t numShapes, shapeAnalysisParameters params);
 //==============================================================================
-void pca_analysis(dataType ** dtaMeanShape, dataType *** eigvectors, dataType ** eigvalues, int * princomp, Shapes *shapes, size_t height, size_t length, size_t width, size_t numShapes, dataType pca_Threshold);
+void pca_analysis(dataType ** dtaMeanShape, dataType *** eigvectors, dataType ** eigvalues, size_t * princomp, Shapes *shapes, size_t height, size_t length, size_t width, size_t numShapes, dataType pca_Threshold);
 //==============================================================================
 // Procrustates distance i s a least square type shape metric that requires two aligned shapes with one-to-one point correspondence
 // Two shapes are equivalent if one can be obtained from the other by translating, scaling and rotating
